Add Contact as a superclass of Address, Phone, and Email
