+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,47 @@
+package seedu.addressbook.data.person;
+
+import seedu.addressbook.data.exception.IllegalValueException;
+
+/**
+ * Represents a Person's email in the address book.
+ * Guarantees: immutable; is valid as declared in {@link #isValidEmail(String)}
+ */
+public class Contact {
+
+    public String value;
+    private boolean isPrivate;
+
+    public Contact() {
+
+    }
+
+    protected void setValue(String value) {
+        this.value = value;
+    }
+
+    protected void setPrivacy(boolean privacy) {
+        this.isPrivate = privacy;
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Contact // instanceof handles nulls
+                && this.value.equals(((Contact) other).value)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
